requirements.txt
pip~=23.3
PyYAML~=6.0.2
pyserial~=3.5
thingspeak~=1.0.0
requests~=2.32.3
proxmoxer~=2.1.0
Flask~=3.1.0